requirements.txt
requests>=2.31.0
feedparser==6.0.11
python-dateutil>=2.8.2
pyyaml>=6.0
openai>=1.0.0
